refactor separable reduced hint
