+++ sympy/solvers/ode/ode.py
@@ -1056,6 +1056,7 @@ def classify_ode(eq, func=None, dict=False, ics=None, *, prep=True, xi=None, eta
         SecondNonlinearAutonomousConserved: ('2nd_nonlinear_autonomous_conserved',),
         Liouville: ('Liouville',),
         Separable: ('separable',),
+        SeparableReduced: ('separable_reduced',),
     }
     for solvercls in solvers:
         solver = solvercls(ode)
@@ -1123,10 +1124,8 @@ def classify_ode(eq, func=None, dict=False, ics=None, *, prep=True, xi=None, eta
             r['d'] = d
             r['e'] = e
             r['y'] = y
-            # print(r)
             r[d] = num.subs(f(x), y)
             r[e] = den.subs(f(x), y)
-            # print(r,"new")
 
             ## Separable Case: y' == P(y)*Q(x)
             r[d] = separatevars(r[d])
@@ -1182,68 +1181,6 @@ def classify_ode(eq, func=None, dict=False, ics=None, *, prep=True, xi=None, eta
                                 'd': d, 'e': e, 'y': y})
                             matching_hints["linear_coefficients"] = r2
                             matching_hints["linear_coefficients_Integral"] = r2
-
-            ## Equation of the form y' + (y/x)*H(x^n*y) = 0
-            # that can be reduced to separable form
-
-            factor = simplify(x/f(x)*num/den)
-
-            # Try representing factor in terms of x^n*y
-            # where n is lowest power of x in factor;
-            # first remove terms like sqrt(2)*3 from factor.atoms(Mul)
-            num, dem = factor.as_numer_denom()
-            num = expand(num)
-            dem = expand(dem)
-            def _degree(expr, x):
-                # Made this function to calculate the degree of
-                # x in an expression. If expr will be of form
-                # x**p*y, (wheare p can be variables/rationals) then it
-                # will return p.
-                for val in expr:
-                    if val.has(x):
-                        if isinstance(val, Pow) and val.as_base_exp()[0] == x:
-                            return (val.as_base_exp()[1])
-                        elif val == x:
-                            return (val.as_base_exp()[1])
-                        else:
-                            return _degree(val.args, x)
-                return 0
-
-            def _powers(expr):
-                # this function will return all the different relative power of x w.r.t f(x).
-                # expr = x**p * f(x)**q then it will return {p/q}.
-                pows = set()
-                if isinstance(expr, Add):
-                    exprs = expr.atoms(Add)
-                elif isinstance(expr, Mul):
-                    exprs = expr.atoms(Mul)
-                elif isinstance(expr, Pow):
-                    exprs = expr.atoms(Pow)
-                else:
-                    exprs = {expr}
-
-                for arg in exprs:
-                    if arg.has(x):
-                        _, u = arg.as_independent(x, f(x))
-                        pow = _degree((u.subs(f(x), y), ), x)/_degree((u.subs(f(x), y), ), y)
-                        pows.add(pow)
-                return pows
-
-            pows = _powers(num)
-            pows.update(_powers(dem))
-            pows = list(pows)
-            if(len(pows)==1) and pows[0]!=zoo:
-                t = Dummy('t')
-                r2 = {'t': t}
-                num = num.subs(x**pows[0]*f(x), t)
-                dem = dem.subs(x**pows[0]*f(x), t)
-                test = num/dem
-                free = test.free_symbols
-                if len(free) == 1 and free.pop() == t:
-                    r2.update({'power' : pows[0], 'u' : test})
-                    matching_hints['separable_reduced'] = r2
-                    matching_hints["separable_reduced_Integral"] = r2
-
 
     elif order == 2:
         # Homogeneous second order differential equation of the form
@@ -4186,82 +4123,6 @@ def ode_linear_coefficients(eq, func, order, match):
     return ode_1st_homogeneous_coeff_best(eq, func, order, match)
 
 
-def ode_separable_reduced(eq, func, order, match):
-    r"""
-    Solves a differential equation that can be reduced to the separable form.
-
-    The general form of this equation is
-
-    .. math:: y' + (y/x) H(x^n y) = 0\text{}.
-
-    This can be solved by substituting `u(y) = x^n y`.  The equation then
-    reduces to the separable form `\frac{u'}{u (\mathrm{power} - H(u))} -
-    \frac{1}{x} = 0`.
-
-    The general solution is:
-
-        >>> from sympy import Function, dsolve, pprint
-        >>> from sympy.abc import x, n
-        >>> f, g = map(Function, ['f', 'g'])
-        >>> genform = f(x).diff(x) + (f(x)/x)*g(x**n*f(x))
-        >>> pprint(genform)
-                         / n     \
-        d          f(x)*g\x *f(x)/
-        --(f(x)) + ---------------
-        dx                x
-        >>> pprint(dsolve(genform, hint='separable_reduced'))
-         n
-        x *f(x)
-          /
-         |
-         |         1
-         |    ------------ dy = C1 + log(x)
-         |    y*(n - g(y))
-         |
-         /
-
-    See Also
-    ========
-    :meth:`sympy.solvers.ode.single.Separable`
-
-    Examples
-    ========
-
-    >>> from sympy import Function, pprint
-    >>> from sympy.solvers.ode.ode import dsolve
-    >>> from sympy.abc import x
-    >>> f = Function('f')
-    >>> d = f(x).diff(x)
-    >>> eq = (x - x**2*f(x))*d - f(x)
-    >>> dsolve(eq, hint='separable_reduced')
-    [Eq(f(x), (1 - sqrt(C1*x**2 + 1))/x), Eq(f(x), (sqrt(C1*x**2 + 1) + 1)/x)]
-    >>> pprint(dsolve(eq, hint='separable_reduced'))
-                   ___________            ___________
-                  /     2                /     2
-            1 - \/  C1*x  + 1          \/  C1*x  + 1  + 1
-    [f(x) = ------------------, f(x) = ------------------]
-                    x                          x
-
-    References
-    ==========
-
-    - Joel Moses, "Symbolic Integration - The Stormy Decade", Communications
-      of the ACM, Volume 14, Number 8, August 1971, pp. 558
-    """
-
-    # Arguments are passed in a way so that they are coherent with the
-    # ode_separable function
-    x = func.args[0]
-    f = func.func
-    y = Dummy('y')
-    u = match['u'].subs(match['t'], y)
-    ycoeff = 1/(y*(match['power'] - u))
-    m1 = {y: 1, x: -1/x, 'coeff': 1}
-    m2 = {y: ycoeff, x: 1, 'coeff': 1}
-    r = {'m1': m1, 'm2': m2, 'y': y, 'hint': x**match['power']*f(x)}
-    return ode_separable(eq, func, order, r)
-
-
 def ode_1st_power_series(eq, func, order, match):
     r"""
     The power series solution is a method which gives the Taylor series expansion
@@ -4919,17 +4780,6 @@ def _solve_variation_of_parameters(eq, func, order, match):
         psol = simplify(psol)
         psol = trigsimp(psol, deep=True)
     return Eq(f(x), gsol.rhs + psol)
-
-
-def ode_separable(eq, func, order, match):
-    x = func.args[0]
-    f = func.func
-    C1 = get_numbered_constants(eq, num=1)
-    r = match  # {'m1':m1, 'm2':m2, 'y':y}
-    u = r.get('hint', f(x))  # get u from separable_reduced else get f(x)
-    return Eq(Integral(r['m2']['coeff']*r['m2'][r['y']]/r['m1'][r['y']],
-        (r['y'], None, u)), Integral(-r['m1']['coeff']*r['m1'][x]/
-        r['m2'][x], x) + C1)
 
 
 def checkinfsol(eq, infinitesimals, func=None, order=None):
@@ -6824,4 +6674,4 @@ def _nonlinear_3eq_order1_type5(x, y, z, t, eq):
 #This import is written at the bottom to avoid circular imports.
 from .single import (NthAlgebraic, Factorable, FirstLinear, AlmostLinear,
         Bernoulli, SingleODEProblem, SingleODESolver, RiccatiSpecial,
-        SecondNonlinearAutonomousConserved, FirstExact, Liouville, Separable)
+        SecondNonlinearAutonomousConserved, FirstExact, Liouville, Separable, SeparableReduced)
